+++ src/main/java/com/woowacamp/soolsool/global/config/CaffeineCacheConfig.java
@@ -1,7 +1,7 @@
 package com.woowacamp.soolsool.global.config;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
-import com.woowacamp.soolsool.global.common.CacheType;
+import com.woowacamp.soolsool.global.infra.CacheType;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
